UnfoldSim

Documentation for UnfoldSim.

UnfoldSim.Component

UnfoldSim.ExperimentDesign

UnfoldSim.convert

UnfoldSim.dims

UnfoldSim.gen_noise

UnfoldSim.gen_noise

UnfoldSim.gen_noise

UnfoldSim.generate

UnfoldSim.padarray

UnfoldSim.simulate

UnfoldSim.simulate_erps

UnfoldSim.Component — Type

ERP Component

UnfoldSim.ExperimentDesign — Type

Experiment Design

UnfoldSim.convert — Method

Function to convert output similar to unfold (data, evts)

UnfoldSim.dims — Method

Returns dimension of experiment design

UnfoldSim.gen_noise — Method

gen_noise(t::RealisticNoise, n::Int)

Generate noise of a given type t and length n

UnfoldSim.gen_noise — Method

gen_noise(t::Union{PinkNoise, RedNoise}, n::Int)

Generate noise of a given type t and length n

UnfoldSim.gen_noise — Method

gen_noise(t::WhiteNoise, n::Int)

Generate noise of a given type t and length n

UnfoldSim.generate — Method

generate(expdesign::ExperimentDesign)

Generates experiment design data frame

UnfoldSim.padarray — Method

Pads array with specified value, length

UnfoldSim.simulate — Method

Simulate eeg data given a simulation design, effect sizes and variances

UnfoldSim.simulate_erps — Method

Simulates erp data given the specified parameters