UnfoldSim.AutoRegressiveNoise — Type

AutoRegressiveNoise <: AbstractNoise

Not implemented

UnfoldSim.ExponentialNoise — Type

ExponentialNoise <: AbstractNoise

Noise with exponential decay in AR spectrum.

noiselevel is used to scale the noise

Warning

With the current implementation we try to get exponential decay over the whole autoregressive (AR) spectrum, which is N samples (the total number of samples in the signal) long. This involves the inversion of a Cholesky matrix of size NxN matrix, which will need lots of RAM for non-trivial problems.

UnfoldSim.LinearModelComponent — Type

LinearModelComponent <: AbstractComponent

A multiple regression component for one subject.

All fields can be named. Works best with SingleSubjectDesign.

Fields

basis::Any: an object, if accessed, provides a 'basis function', e.g. hanning(40)::Vector, this defines the response at a single event. It will be weighted by the model prediction. Future versions will allow for functions, as of v0.3 this is restricted to array-like objects

formula::Any: StatsModels formula object, e.g. @formula 0 ~ 1 + cond (left-hand side must be 0).

β::Vector Vector of betas/coefficients, must fit the formula.

contrasts::Dict (optional): Determines which coding scheme to use for which categorical variables. Default is empty which corresponds to dummy coding. For more information see https://juliastats.org/StatsModels.jl/stable/contrasts.

Examples

julia> LinearModelComponent(;
           basis = hanning(40),
           formula = @formula(0 ~ 1 + cond),
           β = [1., 2.],
           contrasts = Dict(:cond => EffectsCoding())
       )
LinearModelComponent
  basis: Array{Float64}((40,)) [0.0, 0.006474868681043577, 0.02573177902642726, 0.0572719871733951, 0.10027861829824952, 0.1536378232452003, 0.21596762663442215, 0.28565371929847283, 0.3608912680417737, 0.43973165987233853  …  0.43973165987233853, 0.3608912680417737, 0.28565371929847283, 0.21596762663442215, 0.1536378232452003, 0.10027861829824952, 0.0572719871733951, 0.02573177902642726, 0.006474868681043577, 0.0]
  formula: StatsModels.FormulaTerm{StatsModels.ConstantTerm{Int64}, Tuple{StatsModels.ConstantTerm{Int64}, StatsModels.Term}}
  β: Array{Float64}((2,)) [1.0, 2.0]
  contrasts: Dict{Symbol, EffectsCoding}

See also MixedModelComponent, MultichannelComponent.

UnfoldSim.LogNormalOnset — Type

@with_kw struct LogNormalOnset <: AbstractOnset

Log-normal inter-event distances using the Distributions.jl truncated LogNormal distribution.

Be careful with large μ and σ values, as they are on logscale. σ>8 can quickly give you out-of-memory sized signals!

UnfoldSim.MixedModelComponent — Type

MixedModelComponent <: AbstractComponent

A component that adds a hierarchical relation between parameters according to a Linear Mixed Model (LMM) defined via MixedModels.jl.

All fields can be named. Works best with MultiSubjectDesign.

Fields

basis::Any: an object, if accessed, provides a 'basis function', e.g. hanning(40)::Vector, this defines the response at a single event. It will be weighted by the model prediction. Future versions will allow for functions, as of v0.3 this is restricted to array-like objects

formula::Any: Formula-object in the style of MixedModels.jl e.g. @formula 0 ~ 1 + cond + (1|subject). The left-hand side is ignored.

β::Vector Vector of betas (fixed effects), must fit the formula.

σs::Dict Dict of random effect variances, e.g. Dict(:subject => [0.5, 0.4]) or to specify correlation matrix Dict(:subject=>[0.5,0.4,I(2,2)],...). Technically, this will be passed to the MixedModels.jl create_re function, which creates the θ matrices.

contrasts::Dict (optional): Dict in the style of MixedModels.jl. Determines which coding scheme to use for which categorical variables. Default is empty which corresponds to dummy coding. For more information see https://juliastats.org/StatsModels.jl/stable/contrasts.

Examples

julia> MixedModelComponent(;
           basis = hanning(40),
           formula = @formula(0 ~ 1 + cond + (1 + cond|subject)),
           β = [1., 2.],
           σs= Dict(:subject => [0.5, 0.4]),
           contrasts=Dict(:cond => EffectsCoding())
       )
MixedModelComponent
  basis: Array{Float64}((40,)) [0.0, 0.006474868681043577, 0.02573177902642726, 0.0572719871733951, 0.10027861829824952, 0.1536378232452003, 0.21596762663442215, 0.28565371929847283, 0.3608912680417737, 0.43973165987233853  …  0.43973165987233853, 0.3608912680417737, 0.28565371929847283, 0.21596762663442215, 0.1536378232452003, 0.10027861829824952, 0.0572719871733951, 0.02573177902642726, 0.006474868681043577, 0.0]
  formula: StatsModels.FormulaTerm{StatsModels.ConstantTerm{Int64}, Tuple{StatsModels.ConstantTerm{Int64}, StatsModels.Term, StatsModels.FunctionTerm{typeof(|), Vector{StatsModels.AbstractTerm}}}}
  β: Array{Float64}((2,)) [1.0, 2.0]
  σs: Dict{Symbol, Vector{Float64}}
  contrasts: Dict{Symbol, EffectsCoding}

See also LinearModelComponent, MultichannelComponent.

UnfoldSim.MultiSubjectDesign — Type

MultiSubjectDesign <: AbstractDesign

A type for specifying the experimental design for multiple subjects (based on the given random-effects structure).

Fields

n_subjects::Int -> number of subjects

n_items::Int -> number of items (sometimes ≈trials)

subjects_between = Dict{Symbol,Vector} -> effects between subjects, e.g. young vs old

items_between = Dict{Symbol,Vector} -> effects between items, e.g. natural vs artificial images, (but shown to all subjects if not specified also in subjects_between)

both_within = Dict{Symbol,Vector} -> effects completly crossed

event_order_function = (rng, x) -> x; # can be used to sort, or e.g. (rng, x) -> shuffle(rng, x) (or shorter just event_order_function = shuffle)

Tip: Check the resulting dataframe using the generate_events function.

Example

# declaring same condition both sub-between and item-between results in a full between subject/item design
design = MultiSubjectDesign(;
		n_items = 10,
		n_subjects = 30,
		subjects_between = Dict(:cond => ["levelA", "levelB"]),
		items_between = Dict(:cond => ["levelA", "levelB"]),
		);

See also SingleSubjectDesign, RepeatDesign

UnfoldSim.MultichannelComponent — Type

MultichannelComponent <: AbstractComponent

Projects a AbstractComponent to multiple "channels" via the projection vector.

Optionally, noise can be added to the source prior to projection. By default a MultichannelComponent can be constructed using one of the following options for projection:

projection::AbstractVector: Directly pass a custom projection vector.

projection::Pair{<:AbstractHeadmodel,String}: Generate a projection vector by specifying which headmodel to use and which sources should be active.

Fields

component::AbstractComponent: The component that should be projected to the sensors.

projection::AbstractVector or projection::Pair{<:AbstractHeadmodel,String}: Vector p that projects the (source) component c[t] (where t is time) to the sensors s. The length of p equals the number of sensors s. Typically, it is a slice of the leadfield matrix. out[s,t] = p[s]*c[t].

noise::AbstractNoise (optional): Noise added in the source space. Default is NoNoise.

Examples

# Variant 1: Specify the projection vector manually
julia> c1 = LinearModelComponent(; basis = p100(), formula = @formula(0 ~ 1), β = [1]);

julia> mc1 = UnfoldSim.MultichannelComponent(c, [1, 2, -1, 3, 5, 2.3, 1])
MultichannelComponent
  component: LinearModelComponent
  projection: Array{Float64}((7,)) [1.0, 2.0, -1.0, 3.0, 5.0, 2.3, 1.0]
  noise: NoNoise NoNoise()

# Variant 2: Use a headmodel and specify a source
julia> c2 = LinearModelComponent(; basis = p300(), formula = @formula(0 ~ 1), β = [1]);

julia> hart = headmodel(type = "hartmut");
Please cite: HArtMuT: Harmening Nils, Klug Marius, Gramann Klaus and Miklody Daniel - 10.1088/1741-2552/aca8ce

julia> mc2 = UnfoldSim.MultichannelComponent(c2, hart => "Right Occipital Pole")
MultichannelComponent
  component: LinearModelComponent
  projection: Array{Float64}((227,)) [-0.03461859471337842, -0.04321094803502425, 0.0037088347968313525, -0.014722528968861278, -0.0234889834534478, 0.02731807504242923, 0.038863688452528036, 0.1190531258070562, -0.09956890221613562, -0.0867729334438599  …  0.37435404409695094, -0.020863789022627935, 0.25627478723535513, -0.05777985212119245, 0.37104376432271147, -0.19446620423767172, 0.2590764703721097, -0.12923837607416555, 0.1732886690359311, 0.4703016561960567]
  noise: NoNoise NoNoise()

See also LinearModelComponent, MixedModelComponent.

UnfoldSim.NoNoise — Type

NoNoise <: AbstractNoise

Return zeros instead of noise.

UnfoldSim.NoOnset — Type

struct NoOnset <: AbstractOnset end

In the case that the user directly wants no overlap to be simulated (=> epoched data).

UnfoldSim.PinkNoise — Type

PinkNoise <: AbstractNoise

Generate Pink Noise using the SignalAnalysis.jl implementation.

noiselevel is used to scale the noise

UnfoldSim.RealisticNoise — Type

RealisticNoise <: AbstractNoise

Not implemented - planned to use Artifacts.jl to provide real EEG data to add.

UnfoldSim.RedNoise — Type

RedNoise <: AbstractNoise

Generate Red Noise using the SignalAnalysis.jl implementation.

noiselevel is used to scale the noise

UnfoldSim.RepeatDesign — Type

RepeatDesign{T} <: AbstractDesign

Repeat a design DataFrame multiple times to mimick repeatedly recorded trials.

designOnce = MultiSubjectDesign(;
		n_items=2,
		n_subjects = 2,
		subjects_between =Dict(:cond=>["levelA","levelB"]),
		items_between =Dict(:cond=>["levelA","levelB"]),
		);

design = RepeatDesign(designOnce,4);

See also SingleSubjectDesign, MultiSubjectDesign

UnfoldSim.SingleSubjectDesign — Type

SingleSubjectDesign <: AbstractDesign

A type for specifying the experimental for a single subject (based on the given conditions).

Fields

conditions = Dict{Symbol,Vector}of conditions, e.g.Dict(:A=>["asmall","abig"],:B=>["btiny","blarge"])`

event_order_function = (rng::AbstractRNG,x::DataFrame)->x; # can be used to sort by specifying sort, or shuffling by providing shuffle, or custom functions following the interface (rng,x)->my_shuffle(rng,x)

Number of trials / rows in generate_events([rng, ]design) depend on the full factorial of your conditions.

To increase the number of repetitions simply use RepeatDesign(SingleSubjectDesign(...),5)

If conditions are omitted (or set to nothing), a single trial is simulated with a column :dummy and content :dummy - this is for convenience.

Tip: Check the resulting dataframe using the generate_events function.

Examples

design = SingleSubjectDesign(;
    conditions = Dict(
        :stimulus_type => ["natural", "artificial"],
        :contrast_level => range(0, 1, length = 5),
);

See also MultiSubjectDesign, RepeatDesign

UnfoldSim.UniformOnset — Type

struct UniformOnset <: AbstractOnset

Provide a Uniform Distribution of the inter-event-distances. width is the width of the uniform distribution (=> the jitter). Since the lower bound is 0, width is also the upper bound. offset is the minimal distance. The maximal distance is offset + width.

UnfoldSim.WhiteNoise — Type

WhiteNoise <: WhiteNoise
noiselevel = 1
imfilter = 0

Generate White Noise using randn - thus Gaussian noise. noiselevel is used to scale the noise

Using imfilter > 0 it is possible to smooth the noise using Image.imfilter.

Base.size — Method

Returns dimension of experiment design

DSP.Windows.hanning — Method

hanning(duration, offset, sfreq)

Generate a (potentially shifted) hanning window with a certain duration.

Note: This function extends the DSP.hanning function using multiple dispatch.

Arguments

duration: in s.

offset: in s, defines hanning peak i.e. shift of the hanning window.

sfreq: Sampling rate in Hz.

Returns

Vector: Contains a shifted (i.e. zero-padded) hanning window.

Examples

julia> UnfoldSim.hanning(0.1, 0.3, 100)
25-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 0.0
 ⋮
 0.9698463103929542
 0.75
 0.41317591116653485
 0.11697777844051105
 0.0

UnfoldSim.PuRF — Method

PuRF(; n = 10.1, tmax = 0.93, sfreq = 100)

Default generator for PuRF Pupil Response Function. The canonical PRF is a gamma function and implemented according to Denison 2020 equation (2) going back to Hoeks & Levelt, 1993.

The pupil response is evaluated at t = 0:1/sfreq:3*tmax. The response is normalized by the peak-maximum at tmax, thus typically a pupil-response of 1 is returned (disregarding numerical issues).

Keyword arguments:

n = 10.1: shape parameter

tmax = 0.93: peak maximum

sfreq = 100: sampling frequency

Returns

Vector: canonical pupil response with length(0:1/sfreq:3*tmax) entries.

Examples

julia> PuRF(; n = 5)
280-element Vector{Float64}:
 0.0
 2.0216617815131253e-8
 6.130689396024061e-7
 4.4118063684811444e-6
 1.761817666835793e-5
 ⋮
 0.012726253506722554
 0.012280989091455786
 0.011850525657416842
 0.011434405338911133
 0.011032182932283816

UnfoldSim.add_noise! — Method

add_noise!(rng, noisetype::AbstractNoise, signal)

Generate and add noise to a data matrix. Assumes that the signal can be linearized, that is, that the noise is stationary

UnfoldSim.add_responses! — Method

add_responses!(signal, responses::Vector, e, s, tvec, erpvec)
add_responses!(signal, responses::Matrix, e, s, tvec, erpvec)
add_responses!(signal, responses::AbstractArray, e, s, tvec, erpvec)

Helper function to add inplace the responses to the signal, but for both 2D (1 channel) and 3D (X channel case).

UnfoldSim.apply_event_order_function — Method

apply_event_order_function(fun::Function, rng::AbstractRNG, events::DataFrame)

apply fun(rng,events), raise an error if function is wrongly defined. Convenience function to not repeat the error handling at multiple places.

UnfoldSim.closest_src — Method

closest_src(coords_list::AbstractVector{<:AbstractVector}, pos)
closest_src(coords::Vector{<:Real}, pos)

Takes an array of 'm' target coordinate vector (size 3) (or vector of vectors) and a matrix (n-by-3) of all available positions, and returns an array of size 'm' containing the indices of the respective items in 'pos' that are nearest to each of the target coordinates.

UnfoldSim.closest_src — Method

closest_src(head::Hartmut,label::String)

Returns src-ix of the Headmodel Hartmut which is closest to the average of the label.

Important

We use the average in eucledean space, but the cortex is a curved surface. In most cases they will not overlap. Ideally we would calculate the average on the surface, but this is a bit more complex to do (you'd need to calculate the vertices etc.)

hartmut = headmodel()
pos = closest_src(hartmut=>"Left Middle Temporal Gyrus, posterior division")

UnfoldSim.convert — Method

Obsolete - # TODO: Transfer function to Unfold.jl

Function to convert output similar to unfold (data, events)

UnfoldSim.create_continuous_signal — Method

create_continuous_signal(rng, responses, simulation)

Based on the responses and simulation parameters, simulate onset latencies and add together a continuous signal.

UnfoldSim.epoch — Method

epoch(data::AbstractVector, args...; kwargs...)
epoch(
    data::AbstractArray{T,2},
    events,
    τ::Tuple{Number,Number},
    sfreq;
    eventtime::Symbol = :latency,
) where {T<:Union{Missing,Number}}

Helper function to epoch data.

Adapted from Unfold.jl: https://github.com/unfoldtoolbox/Unfold.jl/blob/b3a21c2bb7e93d2f45ec64b0197f4663a6d7939a/src/utilities.jl#L40

UnfoldSim.generate_events — Method

generate_events([rng::AbstractRNG, ]design::MultiSubjectDesign)

Generate full factorial Dataframe according to MixedModelsSim.jl 's simdat_crossed function.

Note: n_items = you can think of it as trials or better, as stimuli. Note: No condition can be named dv which is used internally in MixedModelsSim / MixedModels as a dummy left-side Afterwards applies design.event_order_function. Could be used to duplicate trials, sort, subselect etc. Finally it sorts by :subject

d = MultiSubjectDesign(; n_subjects = 10, n_items = 20, both_within = Dict(:A => nlevels(5),:B => nlevels(2)))
generate_events(d)

UnfoldSim.generate_events — Method

UnfoldSim.generate_events([rng::AbstractRNG, ]design::RepeatDesign{T})

In a repeated design, iteratively calls the underlying {T} design and concatenates. In case of MultiSubjectDesign, sorts by subject.

UnfoldSim.generate_events — Method

generate_events([rng::AbstractRNG, ]design::SingleSubjectDesign)

Generates full-factorial DataFrame of design.conditions

Afterwards applies design.event_order_function. If conditions is nothing, a single trial is simulated with a column :dummy and content :dummy - this is for convenience.

julia> d = SingleSubjectDesign(;conditions= Dict(:A=>nlevels(5),:B=>nlevels(2))) julia> generate_events(d)

UnfoldSim.hartmut_citation — Method

Returns citation-string for HArtMuT

UnfoldSim.headmodel — Method

Load a headmodel, using Artifacts.jl automatically downloads the required files

Currently only type="hartmut" is implemented

UnfoldSim.hrf — Method

hrf(;
TR = 1,
peak = 6.0,
post_undershoot = 16,
length = 32.0,
peak_width = 1.0,
post_undershoot_width = 1,
amplitude = 6,
shift = 0)

Generate a parameterized BOLD haemodynamic response function (HRF) kernel based on gamma-functions.

Implementation and default parameters were taken from the SPM-toolbox.

Note: TR = 1/sfreq

Keyword arguments

TR = 1: repetition time, 1/sfreq.

length = 32.0: total length of the kernel in seconds.

amplitude = 6: maximal amplitude.

peak = 6.0: peak timing.

peak_width = 1.0: width of the peak.

post_undershoot = 16: post-undershoot timing.

post_undershoot_width = 1: post-undershoot width.

shift = 0: shift the whole HRF.

Returns

Vector: HRF BOLD response.

Examples

julia> hrf()
33-element Vector{Float64}:
  0.0
  0.0007715994433635659
  0.019784004131204957
  0.08202939459091822
  0.158157713522699
  ⋮
 -0.0006784790038792572
 -0.00042675060451877775
 -0.000263494738348278
 -0.00015990722628360688
 -9.548780093799345e-5

UnfoldSim.leadfield — Method

Returns the leadfield

UnfoldSim.magnitude — Method

Extracts magnitude of the orientation-including leadfield.

By default uses the orientation specified in the headmodel

Fallback: along the third dimension using norm - the maximal projection

UnfoldSim.magnitude — Method

magnitude(headmodel::Hartmut; type = "perpendicular") = Extract magnitude of 3-orientation-leadfield, type (default: "perpendicular") => uses the provided source-point orientations - otherwise falls back to norm.

UnfoldSim.magnitude — Method

magnitude(lf::AbstractArray{T,3}) where {T<:Real}

If orientation is not specified, returns the maximal magnitude (norm of leadfield).

UnfoldSim.magnitude — Method

magnitude(lf::AbstractArray{T,3}, orientation::AbstractArray{T,2}) where {T<:Real}

Return the magnitude along an orientation of the leadfield.

UnfoldSim.maxlength — Method

maxlength(c::Vector{<:AbstractComponent}) = maximum(length.(c))

Return the maximum of the individual components' lengths.

UnfoldSim.n170 — Method

n170(; sfreq = 100)

Generate a Hanning window mimicking an N170 EEG component with a negative (!) peak at 170ms and a width of 150ms.

Keyword arguments

sfreq = 100: Sampling frequency in Hz.

Returns

Vector: Contains a shifted (i.e. zero-padded) hanning window.

Examples

julia> n170(; sfreq = 120)
28-element Vector{Float64}:
 -0.0
 -0.0
 -0.0
 -0.0
 -0.0
  ⋮
 -0.45386582026834904
 -0.2771308221117309
 -0.1304955413896705
 -0.03376388529782215
 -0.0

See also p100, p300, n400, hanning.

UnfoldSim.n400 — Method

n400(; sfreq = 100)

Generate a Hanning window mimicking an N400 EEG component with a negative (!) peak at 400ms and a width of 400ms.

Keyword arguments

sfreq = 100: Sampling frequency in Hz.

Returns

Vector: Contains a shifted (i.e. zero-padded) hanning window.

Examples

julia> n400(; sfreq = 250)
150-element Vector{Float64}:
 -0.0
 -0.0
 -0.0
 -0.0
 -0.0
  ⋮
 -0.016025649301821876
 -0.009035651368646647
 -0.00402259358460233
 -0.0010066617640578368
 -0.0

See also p100, p300, n170, hanning.

UnfoldSim.n_channels — Method

n_channels(c::AbstractComponent)

Return the number of channels for the given component c. By default = 1.

UnfoldSim.n_channels — Method

n_channels(c::MultichannelComponent)

For MultichannelComponent return the length of the projection vector.

UnfoldSim.n_channels — Method

n_channels(c::Vector{<:AbstractComponent})

For a vector of MultichannelComponents, return the number of channels for the first component but assert all are of equal length.

UnfoldSim.p100 — Method

p100(; sfreq = 100)

Generate a Hanning window mimicking a P100 EEG component with a peak at 100ms and a width of 100ms.

Keyword arguments

sfreq = 100: Sampling frequency in Hz.

Returns

Vector: Contains a shifted (i.e. zero-padded) hanning window.

Examples

julia> p100(; sfreq = 200)
30-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 0.0
 ⋮
 0.37725725642960045
 0.22652592093878665
 0.10542974530180327
 0.02709137914968268
 0.0

See also p300, n170, n400, hanning.

UnfoldSim.p300 — Method

p300(; sfreq = 100)

Generate a Hanning window mimicking a P300 EEG component with a peak at 300ms and a width of 300ms.

Keyword arguments

sfreq = 100: Sampling frequency in Hz.

Returns

Vector: Contains a shifted (i.e. zero-padded) hanning window.

Examples

julia> p300(; sfreq = 150)
67-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 0.0
 ⋮
 0.07937323358440934
 0.04518400232274078
 0.02025351319275137
 0.005089279059533658
 0.0

See also p100, n170, n400, hanning.

UnfoldSim.pad_array — Method

Pads array with specified value, length pad_array(arr, len, val)

UnfoldSim.predef_2x2 — Method

predef_2x2(rng::AbstractRNG; kwargs...)

The most used kwargs is: return_epoched = true which returns already epoched data. If you want epoched data without overlap, specify onset = NoOnset() and return_epoched = true

Design

n_items = 100,

n_subjects = 1,

conditions = Dict(:A => ["a_small","a_big"], :B => ["b_tiny","b_large"]),

event_order_function = shuffle,

Component / Signal

signalsize = 100, # Length of simulated hanning window

basis = hanning(signalsize), # The actual "function", signalsize is only used here

β = [1, -0.5, .5, +1], # The parameters

σs = Dict(:subject => [1, 0.5, 0.5, 0.5],:item => [1]), # Only in n_subjects >= 2 case, specifies the random effects

contrasts = Dict(:A => EffectsCoding(), :B => EffectsCoding()) # Effect coding by default

formula = n_subjects == 1 ? @formula(0 ~ 1 + A*B) : @formula(dv ~ 1 + A*B + (A*B|subject) + (1|item)),

Onset

overlap = (0.5,0.2),

onset = UniformOnset(; offset = signalsize * overlap[1], width = signalsize * overlap[2]), # Put offset to 1 for no overlap. put width to 0 for no jitter

Noise

noiselevel = 0.2,

noise = PinkNoise(; noiselevel = noiselevel),

Be careful if you modify nitems with nsubjects = 1, n_items has to be a multiple of 4 (or your equivalent conditions factorial, e.g. all combinations length)

UnfoldSim.predef_eeg — Method

predef_eeg(; kwargs...)
predef_eeg(rng; kwargs...)
predef_eeg(rng, n_subjects; kwargs...)

Generate a P1/N1/P3 complex. In case n_subjects is defined - MixedModelComponents are generated, else LinearModelComponents.

The most used kwargs is: return_epoched=true which returns already epoched data. If you want epoched data without overlap, specify onset = NoOnset() and return_epoched = true

Default parameters:

Design

n_repeats = 100,

event_order_function = shuffle, # random trial order

conditions = Dict(...),

Component / Signal

sfreq = 100,

p1 = (p100(; sfreq = sfreq), @formula(0 ~ 1), [5], Dict()), # P1 amp 5, no effects

n1 = (n170(; sfreq = sfreq), @formula(0 ~ 1 + condition), [5,-3], Dict()), # N1 amp 5, dummy-coded condition effect (levels "car", "face") of -3

p3 = (p300(; sfreq = sfreq), @formula(0 ~ 1 + continuous), [5,1], Dict()), # P3 amp 5, continuous effect range [-5,5] with slope 1

Onset

overlap = (0.5,0.2), # offset + width/length of Uniform noise. put offset to 1 for no overlap. put width to 0 for no jitter

onset = UniformOnset(; offset = sfreq * 0.5 * overlap[1], width = sfreq * 0.5 * overlap[2]),

Noise

noiselevel = 0.2,

noise = PinkNoise(; noiselevel = noiselevel),

UnfoldSim.simulate — Function

simulate(
rng::AbstractRNG,
design::AbstractDesign,
signal,
onset::AbstractOnset,
noise::AbstractNoise = NoNoise();
return_epoched = false,
)

simulate(
design::AbstractDesign,
signal,
onset::AbstractOnset,
noise::AbstractNoise = NoNoise();
return_epoched = false,
)

Return continuous or epoched signal, given Design, [Array of] Component, Onset and optional [Noise] and [RNG]. Main simulation function.

Arguments

design::AbstractDesign: Desired experimental design.

signal: Component for the desired signal.

onset::AbstractOnset: Desired onset.

noise::AbstractNoise = NoNoise() (optional): Desired noise.

rng::AbstractRNG (optional): Random number generator, important to ensure reproducibility.

Keyword arguments

return_epoched = false: Skip the Onset-calculation and conversion to continuous data and return the epoched data directly (see also Notes below).

Returns

signal : Generated signal. Depending on the design, on the components and on return_epoched, the output can be a 1-D, 2-D, 3-D or 4-D Array. For example, a 4-D Array would have the dimensions channels x time x trials x subjects.

events: Generated events.

Examples

Adapted from the quickstart tutorial in the UnfoldSim docs.

julia> using UnfoldSim

julia> using Random # to get an RNG

julia> design =
    SingleSubjectDesign(; conditions = Dict(:cond_A => ["level_A", "level_B"])) |>
    x -> RepeatDesign(x, 10);

julia> signal = LinearModelComponent(; 
    basis = [0, 0, 0, 0.5, 1, 1, 0.5, 0, 0],
    formula = @formula(0 ~ 1 + cond_A),
    β = [1, 0.5],
);

julia> onset = UniformOnset(; width = 20, offset = 4);

julia> noise = PinkNoise(; noiselevel = 0.2);

julia> data, events = simulate(MersenneTwister(1), design, signal, onset, noise)
([-0.045646938524459196, 0.15784406738265955, 0.012640319497460443, 0.026669512219327673, 0.15329144053662508, 0.06412654786607011, -0.16766777448918685, -0.08012027590515228, 0.0020515088981202137, -0.24874482217391175  …  0.24621397283439814, 0.1710771262918883, -0.01527736524528042, 0.7639978745937471, 1.5600315092771557, 1.624219837479329, 1.2889713838347956, 0.26819223928179, 0.16535758767503866, 0.21291936972924855], 20×2 DataFrame
 Row │ cond_A   latency 
     │ String   Int64   
─────┼──────────────────
   1 │ level_A       18
   2 │ level_B       39
   3 │ level_A       45
   4 │ level_B       49
  ⋮  │    ⋮        ⋮
  18 │ level_B      258
  19 │ level_A      281
  20 │ level_B      303

julia> data1, events1 = simulate(design, signal, onset, noise)
┌ Warning: No random generator defined, used the default (`Random.MersenneTwister(1)`) with a fixed seed. This will always return the same results and the user is strongly encouraged to provide their own random generator!

Notes

Some remarks to how the noise is added:

If return_epoched = true and onset = NoOnset() the noise is added to the epoched data matrix.

If onset is not NoOnset, a continuous signal is created and the noise is added to this i.e. this means that the noise won't be the same as in the onset = NoOnset() case even if return_epoched = true.

The case return_epoched = false and onset = NoOnset() is not possible and therefore covered by an assert statement.

UnfoldSim.simulate_and_add! — Method

simulate_and_add!(epoch_data::AbstractMatrix, c, simulation, rng)
simulate_and_add!(epoch_data::AbstractArray, c, simulation, rng)

Helper function to call simulate_component and add it to a provided Array epoch_data.

UnfoldSim.simulate_component — Method

simulate_component(rng, c::AbstractComponent, simulation::Simulation)

By default call simulate_component with (rng, c::Abstractcomponent, design::AbstractDesign) instead of the whole simulation. This function exist solely to provide a "hook" if for a custom component something else than the design is necessary, e.g. a dependency on the onsets, noise or similar.

UnfoldSim.simulate_component — Method

simulate_component(rng, c::LinearModelComponent, design::AbstractDesign)

Generate a linear model design matrix, weight it by the coefficients c.β and multiply the result with the given basis vector.

Returns

Matrix{Float64}: Simulated component for each event in the events dataframe. The output dimensions are length(c.basis) x length(design).

Examples

julia> design = SingleSubjectDesign(; conditions = Dict(:cond => ["natural", "artificial"]));

julia> c = UnfoldSim.LinearModelComponent([0, 1, 1, 0], @formula(0 ~ 1 + cond), [1, 2], Dict());

julia> using StableRNGs

julia> simulate_component(StableRNG(1), c, design)
4×2 Matrix{Float64}:
 0.0  0.0
 3.0  1.0
 3.0  1.0
 0.0  0.0

UnfoldSim.simulate_component — Method

simulate_component(rng, c::MixedModelComponent, design::AbstractDesign, return_parameters = false)

Generate a MixedModel and simulate data according to the given paraemters c.β and c.σs.

Keyword arguments

return_parameters::Bool = false: Can be used to return the per-event parameters used to weight the basis function. Sometimes useful to inspect what is simulated.

Returns

Matrix{Float64}: Simulated component for each event in the events dataframe. The output dimensions are length(c.basis) x length(design).

Notes

MixedModels/Sim does not allow simulation of data without white noise of the residuals. Because we want our own noise, we use the following trick to remove the MixedModels-Noise:

Practically, we upscale the specified σs by factor 10000, and request a white-noise-level of σ = 0.0001. Internally in MixedModels/Sim, σs are relative to σ, and thus are normalized correctly, while keeping the noise 10000 times smaller than the random effects

We cannot exclude that this trick runs into strange numerical issues if the random effect σs are very large compared to the fixed effects.

Currently, it is not possible to use a different basis for fixed and random effects. If this is needed, some code-scaffold is available but commented out at the moment and requires a bit of implementation work.

Examples

julia> design = MultiSubjectDesign(; n_subjects = 2, n_items = 6, items_between = Dict(:cond => ["A", "B"]));

julia> c = UnfoldSim.MixedModelComponent([0, 1, 1, 0], @formula(0 ~ 1 + cond + (1|subject)), [1, 2], Dict(:subject => [2],), Dict());

julia> using StableRNGs

julia> simulate_component(StableRNG(1), c, design)
4×12 Matrix{Float64}:
 -0.0      -0.0       -0.0      -0.0       -0.0     -0.0       0.0       0.0      0.0       0.0      0.0       0.0
 -2.70645  -0.706388  -2.70632  -0.706482  -2.7066  -0.706424  0.325722  2.32569  0.325627  2.32564  0.325468  2.32565
 -2.70645  -0.706388  -2.70632  -0.706482  -2.7066  -0.706424  0.325722  2.32569  0.325627  2.32564  0.325468  2.32565
 -0.0      -0.0       -0.0      -0.0       -0.0     -0.0       0.0       0.0      0.0       0.0      0.0       0.0

julia> simulate_component(StableRNG(1), c, design, return_parameters = true)
1×12 Matrix{Float64}:
 -2.70645  -0.706388  -2.70632  -0.706482  -2.7066  -0.706424  0.325722  2.32569  0.325627  2.32564  0.325468  2.32565

UnfoldSim.simulate_component — Method

simulate_component(rng, c::MultichannelComponent, design::AbstractDesign)

Return the projection of a MultichannelComponent c from "source" to "sensor" space.

Returns

Array{Float64,3}: Projected simulated component for each event in the events dataframe. The output dimensions are length(c.projection) x length(c.basis) x length(design).

Examples

julia> design = SingleSubjectDesign(; conditions = Dict(:cond => ["natural", "artificial"]));

julia> c = LinearModelComponent(; basis = p100(), formula = @formula(0 ~ 1 + cond), β = [1, 0.5]);
julia> mc = UnfoldSim.MultichannelComponent(c, [1, 2, -1, 3, 5, 2.3, 1], PinkNoise());

julia> using StableRNGs

julia> simulate_component(StableRNG(1), mc, design)
7×15×2 Array{Float64, 3}:
[:, :, 1] =
  0.859626   1.16574   0.959523   0.757522   1.17639   1.65156   1.3742    1.76706   2.76971   2.0306    1.17429   1.00922   1.09519   0.754659   2.25662
  1.71925    2.33147   1.91905    1.51504    2.35278   3.30312   2.7484    3.53412   5.53942   4.0612    2.34858   2.01845   2.19039   1.50932    4.51324
 -0.859626  -1.16574  -0.959523  -0.757522  -1.17639  -1.65156  -1.3742   -1.76706  -2.76971  -2.0306   -1.17429  -1.00922  -1.09519  -0.754659  -2.25662
  2.57888    3.49721   2.87857    2.27257    3.52917   4.95469   4.1226    5.30118   8.30913   6.09179   3.52287   3.02767   3.28558   2.26398    6.76985
  4.29813    5.82868   4.79761    3.78761    5.88194   8.25781   6.871     8.8353   13.8485   10.153     5.87145   5.04612   5.47597   3.7733    11.2831
  1.97714    2.68119   2.2069     1.7423     2.70569   3.79859   3.16066   4.06424   6.37033   4.67037   2.70087   2.32121   2.51894   1.73572    5.19022
  0.859626   1.16574   0.959523   0.757522   1.17639   1.65156   1.3742    1.76706   2.76971   2.0306    1.17429   1.00922   1.09519   0.754659   2.25662

[:, :, 2] =
  0.859626   1.16574   0.959523   0.757522   1.17639   1.65156   1.31571   1.56047   2.39471   1.54567   0.689367   0.634223   0.888605   0.69617   2.25662
  1.71925    2.33147   1.91905    1.51504    2.35278   3.30312   2.63142   3.12094   4.78942   3.09135   1.37873    1.26845    1.77721    1.39234   4.51324
 -0.859626  -1.16574  -0.959523  -0.757522  -1.17639  -1.65156  -1.31571  -1.56047  -2.39471  -1.54567  -0.689367  -0.634223  -0.888605  -0.69617  -2.25662
  2.57888    3.49721   2.87857    2.27257    3.52917   4.95469   3.94713   4.68142   7.18413   4.63702   2.0681     1.90267    2.66582    2.08851   6.76985
  4.29813    5.82868   4.79761    3.78761    5.88194   8.25781   6.57855   7.80236  11.9735    7.72837   3.44684    3.17112    4.44303    3.48085  11.2831
  1.97714    2.68119   2.2069     1.7423     2.70569   3.79859   3.02613   3.58909   5.50783   3.55505   1.58554    1.45871    2.04379    1.60119   5.19022
  0.859626   1.16574   0.959523   0.757522   1.17639   1.65156   1.31571   1.56047   2.39471   1.54567   0.689367   0.634223   0.888605   0.69617   2.25662

UnfoldSim.simulate_noise — Method

simulate_noise(rng, t::NoNoise, n::Int)

Return zeros instead of noise.

UnfoldSim.simulate_noise — Method

simulate_noise(rng, t::Union{PinkNoise,RedNoise}, n::Int)

Generate Pink or Red Noise using the SignalAnalysis.jl implementation.

UnfoldSim.simulate_onsets — Method

simulate_onsets(rng, onset::AbstractOnset, simulation::Simulation)

Call simulate_interonset_distances to generate distances between events and then add them up to generate the actual latencies in samples.

main call from simulation

UnfoldSim.simulate_responses — Method

simulate_responses(
    rng,
    components::Vector{<:AbstractComponent},
    simulation::Simulation)

Simulate multiple component responses and accumulate them on a per-event basis.

Returns

epoch_data: Matrix (or Array in the multi-channel case) of combined simulated components. The output dimensions are maxlength(components) x length(simulation.design) for single-channel components and n_channels(components) x maxlength(components) x length(simulation.design) for multi-channel components.

Examples

julia> design = SingleSubjectDesign(; conditions = Dict(:cond => ["natural", "artificial"]));

julia> c1 = LinearModelComponent(; basis = p100(), formula = @formula(0 ~ 1 + cond), β = [1, 0.5]);
julia> c2 = LinearModelComponent(; basis = p300(), formula = @formula(0 ~ 1), β = [2]);

julia> simulation = Simulation(design, [c1, c2], UniformOnset(; width = 0, offset = 30), PinkNoise());

julia> using StableRNGs

julia> simulate_responses(StableRNG(1), [c1, c2], simulation)
45×2 Matrix{Float64}:
 0.0        0.0
 0.0        0.0
 0.0        0.0
 0.0        0.0
 0.0        0.0
 ⋮          
 0.352614   0.352614
 0.203907   0.203907
 0.0924246  0.0924246
 0.0233794  0.0233794
 0.0        0.0

UnfoldSim.weight_σs — Method

weight_σs(σs::Dict, b_σs::Float64, σ_lmm::Float64)

Weight a σs Dict for MixedModels.jl by b_σs, a scaling factor typically from a basis.

Finally scales it again by σ_lmm, as a trick to simulate noise-free LMMs (see MixedModelsComponent)

In the future, we anticipate a function function weight_σs(σs::Dict,b_σs::Dict,σ_lmm::Float64) where each σs entry can be weighted individually by a matching b_σs, but it is not implemented.

Arguments

σs::Dict = a Dict of random effects as output of MixedModels.create_re

b_σs::Float64 = a scaling factor, typically one entry of a basis function from a component

σ_lmm::Float64 = a scaling factor to simulate near-zero noise LMMs

Returns

`NamedTuple` of the weighted random effects