UnfoldSim.LinearModelComponent — Type

A multiple regression component for one subject

basis: an object, if accessed, provides a 'basis-function', e.g. hanning(40), this defines the response at a single event. It will be weighted by the model-prediction

formula: StatsModels Formula-Object @formula 0~1+cond (left side must be 0)

β Vector of betas, must fit the formula

contrasts: Dict. Default is empty, e.g. Dict(:condA=>EffectsCoding())

All arguments can be named, in that case contrasts is optional

Works best with SingleSubjectDesign

LinearModelComponent(;
    basis=hanning(40),
    formula=@formula(0~1+cond),
    β = [1.,2.],
    contrasts=Dict(:cond=>EffectsCoding())
)

UnfoldSim.MixedModelComponent — Type

A component that adds a hierarchical relation between parameters according to a LMM defined via MixedModels.jl

basis: an object, if accessed, provides a 'basis-function', e.g. hanning(40), this defines the response at a single event. It will be weighted by the model-prediction

formula: Formula-Object in the style of MixedModels.jl e.g. @formula dv~1+cond + (1|subject) - left side must be dv

β Vector of betas, must fit the formula

σs Dict of random effect variances, e.g. Dict(:subject=>[0.5,0.4]) or to specify correlationmatrix Dict(:subject=>[0.5,0.4,I(2,2)],...). Technically, this will be passed to MixedModels.jl create_re function, which creates the θ matrices.

contrasts: Dict in the style of MixedModels.jl. Default is empty.

All arguments can be named, in that case contrasts is optional

Works best with MultiSubjectDesign

MixedModelComponent(;
    basis=hanning(40),
    formula=@formula(dv~1+cond+(1+cond|subject)),
    β = [1.,2.],
    σs= Dict(:subject=>[0.5,0.4]),
    contrasts=Dict(:cond=>EffectsCoding())
)

UnfoldSim.MultiSubjectDesign — Type

n_subjects::Int -> number of subjects

n_items::Int -> number of items (sometimes ≈trials)

subjects_between = nothing -> effects between subjects, e.g. young vs old

items_between = nothing -> effects between items, e.g. natural vs artificial images, but shown to all subjects

both_within = nothing -> effects completly crossed

tableModifyFun = x->x; # can be used to sort, or x->shuffle(MersenneTwister(42),x) - be sure to fix/update the rng accordingly!!

tipp: check the resulting dataframe using generate(design)

# declaring same condition both sub-between and item-between results in a full between subject/item design
design = MultiSubjectDesignjectDesign(;
        n_items=10,
		n_subjects = 30,
        subjects_between=Dict(:cond=>["levelA","levelB"]),
		items_between =Dict(:cond=>["levelA","levelB"]),
        );

UnfoldSim.RepeatDesign — Type

repeat a design DataFrame multiple times to mimick repeatedly recorded trials

designOnce = MultiSubjectDesign(;
        n_items=2,
		n_subjects = 2,
        subjects_between =Dict(:cond=>["levelA","levelB"]),
		items_between =Dict(:cond=>["levelA","levelB"]),
        );

design = RepeatDesign(designOnce,4);

UnfoldSim.SingleSubjectDesign — Type

conditions = Dict of conditions, e.g. Dict(:A=>["a_small","a_big"],:B=>["b_tiny","b_large"])

tableModifyFun = x->x; # can be used to sort, or x->shuffle(MersenneTwister(42),x) - be sure to fix/update the rng accordingly!!

Number of trials / rows in generate(design) depend on the full factorial of your conditions.

To increase the number of repetitions simply use RepeatDesign(SingleSubjectDesign(...),5)

tipp: check the resulting dataframe using generate(design)

Base.size — Method

Returns dimension of experiment design

DSP.Windows.hanning — Method

generate a hanning window

duration: in s offset: in s, defines hanning peak sfreq: sampling rate in Hz

UnfoldSim.convert — Method

Function to convert output similar to unfold (data, evts)

UnfoldSim.gen_noise — Method

gen_noise(t::RealisticNoise, n::Int)

Generate noise of a given type t and length n

UnfoldSim.gen_noise — Method

gen_noise(t::Union{PinkNoise, RedNoise}, n::Int)

Generate noise of a given type t and length n

UnfoldSim.gen_noise — Method

gen_noise(t::WhiteNoise, n::Int)

Generate noise of a given type t and length n

UnfoldSim.generate — Method

Generates full factorial Dataframe according to MixedModelsSim.jl 's simdatcrossed function Note: nitems = you can think of it as trials or better, as stimuli

Note: No condition can be named dv which is used internally in MixedModelsSim / MixedModels as a dummy left-side

Afterwards applies expdesign.tableModifyFun. Could be used to duplicate trials, sort, subselect etc.

Finally it sorts by :subject

julia> d = MultiSubjectDesign(;nsubjects = 10,nitems=20,both_within= Dict(:A=>nlevels(5),:B=>nlevels(2))) julia> generate(d)

UnfoldSim.generate — Method

Generates full-factorial DataFrame of expdesign.conditions

Afterwards applies expdesign.tableModifyFun.

julia> d = SingleSubjectDesign(;conditions= Dict(:A=>nlevels(5),:B=>nlevels(2))) julia> generate(d)

UnfoldSim.hrf — Method

Generate a HRF kernel.

TR = 1/sfreq default parameters taken from SPM

Code adapted from Unfold.jl

UnfoldSim.padarray — Method

Pads array with specified value, length padarray(arr, len, val)

UnfoldSim.predef_2x2 — Method

todo

Careful if you modify nitems with nsubjects = 1, n_items has to be a multiple of 4 (or your equivalent conditions factorial, e.g. all combinations length)

UnfoldSim.simulate — Method

Simulate eeg data given a simulation design, effect sizes and variances

make use of return_epoched=true to skip the Onset-calculation + conversion to continuous data and get the epoched data directly

UnfoldSim.simulate — Method

by default call simulate with ::Abstractcomponent,::AbstractDesign`, but allow for custom types

making use of other information in simulation

UnfoldSim.simulate — Method

simulate a linearModel

julia> c = UnfoldSim.LinearModelComponent([0,1,1,0],@formula(0~1+cond),[1,2],Dict()) julia> design = MultiSubjectDesign(;nsubjects=2,nitems=50,item_between=(;:cond=>["A","B"])) julia> simulate(StableRNG(1),c,design)

UnfoldSim.simulate — Method

simulate MixedModelComponent

julia> design = MultiSubjectDesign(;nsubjects=2,nitems=50,item_between=(;:cond=>["A","B"])) julia> c = UnfoldSim.MixedModelComponent([0.,1,1,0],@formula(dv~1+cond+(1|subject)),[1,2],Dict(:subject=>[2],),Dict()) julia> simulate(StableRNG(1),c,design)

UnfoldSim.simulate — Method

Simulates erp data given the specified parameters

UnfoldSim.weight_σs — Method

Weights a σs Dict for MixedModels.jl by a Float64

Finally sales it by σ_lmm, as a trick to simulate noise-free LMMs

I anticipate a function function weight_σs(σs::Dict,b_σs::Dict,σ_lmm::Float64) where each σs entry can be weighted individually