UnfoldSim.LinearModelComponent — Type

A multiple regression component for one subject

basis: an object, if accessed, provides a 'basis-function', e.g. hanning(40), this defines the response at a single event. It will be weighted by the model-prediction

formula: StatsModels Formula-Object @formula 0~1+cond (left side must be 0)

β Vector of betas, must fit the formula

contrasts: Dict. Default is empty, e.g. Dict(:condA=>EffectsCoding())

All arguments can be named, in that case contrasts is optional

Works best with SingleSubjectDesign

LinearModelComponent(;
    basis=hanning(40),
    formula=@formula(0~1+cond),
    β = [1.,2.],
    contrasts=Dict(:cond=>EffectsCoding())
)

UnfoldSim.MixedModelComponent — Type

A component that adds a hierarchical relation between parameters according to a LMM defined via MixedModels.jl

basis: an object, if accessed, provides a 'basis-function', e.g. hanning(40), this defines the response at a single event. It will be weighted by the model-prediction

formula: Formula-Object in the style of MixedModels.jl e.g. @formula 0~1+cond + (1|subject) - left-handside is ignored

β Vector of betas, must fit the formula

σs Dict of random effect variances, e.g. Dict(:subject=>[0.5,0.4]) or to specify correlationmatrix Dict(:subject=>[0.5,0.4,I(2,2)],...). Technically, this will be passed to MixedModels.jl create_re function, which creates the θ matrices.

contrasts: Dict in the style of MixedModels.jl. Default is empty.

All arguments can be named, in that case contrasts is optional

Works best with MultiSubjectDesign

MixedModelComponent(;
    basis=hanning(40),
    formula=@formula(0~1+cond+(1+cond|subject)),
    β = [1.,2.],
    σs= Dict(:subject=>[0.5,0.4]),
    contrasts=Dict(:cond=>EffectsCoding())
)

UnfoldSim.MultiSubjectDesign — Type

n_subjects::Int -> number of subjects

n_items::Int -> number of items (sometimes ≈trials)

subjects_between = nothing -> effects between subjects, e.g. young vs old

items_between = nothing -> effects between items, e.g. natural vs artificial images, but shown to all subjects

both_within = nothing -> effects completly crossed

tableModifyFun = x->x; # can be used to sort, or x->shuffle(MersenneTwister(42),x) - be sure to fix/update the rng accordingly!!

tipp: check the resulting dataframe using generate(design)

# declaring same condition both sub-between and item-between results in a full between subject/item design
design = MultiSubjectDesign(;
		n_items = 10,
		n_subjects = 30,
		subjects_between = Dict(:cond => ["levelA", "levelB"]),
		items_between = Dict(:cond => ["levelA", "levelB"]),
		);

UnfoldSim.MultichannelComponent — Type

Wrapper for an AbstractComponent to project it to multiple target-channels via projection. optional adds noise to the source prior to projection.

UnfoldSim.RepeatDesign — Type

repeat a design DataFrame multiple times to mimick repeatedly recorded trials

designOnce = MultiSubjectDesign(;
		n_items=2,
		n_subjects = 2,
		subjects_between =Dict(:cond=>["levelA","levelB"]),
		items_between =Dict(:cond=>["levelA","levelB"]),
		);

design = RepeatDesign(designOnce,4);

UnfoldSim.SingleSubjectDesign — Type

conditions = Dict of conditions, e.g. Dict(:A=>["a_small","a_big"],:B=>["b_tiny","b_large"])

tableModifyFun = x->x; # can be used to sort, or x->shuffle(MersenneTwister(42),x) - be sure to fix/update the rng accordingly!!

Number of trials / rows in generate(design) depend on the full factorial of your conditions.

To increase the number of repetitions simply use RepeatDesign(SingleSubjectDesign(...),5)

If conditions are omitted (or set to nothing), a single trial is simulated with a column :dummy and content :dummy - this is for convenience.

tipp: check the resulting dataframe using generate(design)

Base.size — Method

Returns dimension of experiment design

DSP.Windows.hanning — Method

generate a hanning window

duration: in s offset: in s, defines hanning peak sfreq: sampling rate in Hz

UnfoldSim.adderp! — Method

Helper function to add inplace the erps to the EEG, but for both 2D (1 channel) and 3D (X channel case)

UnfoldSim.closest_src — Method

closest_src(coords_list::AbstractVector{<:AbstractVector}, pos)
closest_src(coords::Vector{<:Real}, pos)

Takes an array of 'm' target coordinate vector (size 3) (or vector of vectors) and a matrix (n-by-3) of all available positions, and returns an array of size 'm' containing the indices of the respective items in 'pos' that are nearest to each of the target coordinates.

UnfoldSim.closest_src — Method

closest_src(head::Hartmut,label::String)

Returns src-ix of the Headmodel Hartmut which is closest to the average of the label.

Important

We use the average in eucledean space, but the cortex is a curved surface. In most cases they will not overlap. Ideally we would calculate the average on the surface, but this is a bit more complex to do (you'd need to calculate the vertices etc.)

hartmut = headmodel()
pos = closest_src(hartmut=>"Left Middle Temporal Gyrus, posterior division")

UnfoldSim.convert — Method

Function to convert output similar to unfold (data, evts)

UnfoldSim.gen_noise — Method

gen_noise(t::RealisticNoise, n::Int)

Generate noise of a given type t and length n

UnfoldSim.gen_noise — Method

gen_noise(t::Union{PinkNoise, RedNoise}, n::Int)

Generate noise of a given type t and length n

UnfoldSim.gen_noise — Method

gen_noise(t::WhiteNoise, n::Int)

Generate noise of a given type t and length n

UnfoldSim.generate — Method

Generates full factorial Dataframe according to MixedModelsSim.jl 's simdatcrossed function Note: nitems = you can think of it as trials or better, as stimuli

Note: No condition can be named dv which is used internally in MixedModelsSim / MixedModels as a dummy left-side

Afterwards applies expdesign.tableModifyFun. Could be used to duplicate trials, sort, subselect etc.

Finally it sorts by :subject

julia> d = MultiSubjectDesign(;nsubjects = 10,nitems=20,both_within= Dict(:A=>nlevels(5),:B=>nlevels(2))) julia> generate(d)

UnfoldSim.generate — Method

Generates full-factorial DataFrame of expdesign.conditions

Afterwards applies expdesign.tableModifyFun.

If conditions is nothing, a single trial is simulated with a column :dummy and content :dummy - this is for convenience.

julia> d = SingleSubjectDesign(;conditions= Dict(:A=>nlevels(5),:B=>nlevels(2))) julia> generate(d)

UnfoldSim.hartmut_citation — Method

Returns citation-string for HArtMuT

UnfoldSim.headmodel — Method

Load a headmodel, using Artifacts.jl automatically downloads the required files

Currently only type="hartmut" is implemented

UnfoldSim.hrf — Method

Generate a HRF kernel.

TR = 1/sfreq default parameters taken from SPM

Code adapted from Unfold.jl

UnfoldSim.leadfield — Method

Returns the leadfield

UnfoldSim.magnitude — Method

Extracts magnitude of the orientation-including leadfield.

By default uses the orientation specified in the headmodel

Fallback: along the third dimension using norm - the maximal projection

UnfoldSim.magnitude — Method

Extract magnitude of 3-orientation-leadfield, type (default: "perpendicular") => uses the provided source-point orientations - otherwise falls back to norm

UnfoldSim.n_channels — Method

Returns the number of channels. By default = 1

UnfoldSim.n_channels — Method

for MultichannelComponent returns the length of the projection vector

UnfoldSim.padarray — Method

Pads array with specified value, length padarray(arr, len, val)

UnfoldSim.predef_2x2 — Method

predef_2x2(rng::AbstractRNG;kwargs...)

Most used kwargs is: return_epoched=true to ignore the overlap/onset bits and return already epoched data

design

n_items=100,

n_subjects=1,

conditions = Dict(:A=>["asmall","abig"],:B=>["btiny","blarge"]),

tableModifyFun = x->shuffle(deepcopy(rng),x),

component / signal

signalsize = 100, length of simulated hanning window

basis= hanning(signalsize), the actual "function",signalsize` is only used here

β = [1,-0.5,.5,+1], the parameters

σs = Dict(:subject=>[1,0.5,0.5,0.5],:item=>[1]), - only in n_subjects>=2 case, specifies the random effects

contrasts = Dict(:A=>EffectsCoding(),:B=>EffectsCoding()) - effect coding by default

formula = n_subjects==1 ? @formula(0~1+AB) : @formula(dv~1+AB+(A*B|subject)+(1|item)),

noise

noiselevel = 0.2,

noise = PinkNoise(;noiselevel=noiselevel),

onset

overlap = (0.5,0.2),

onset=UniformOnset(;offset=signalsizeoverlap[1],width=signalsizeoverlap[2]), #put offset to 1 for no overlap. put width to 0 for no jitter

Careful if you modify nitems with nsubjects = 1, n_items has to be a multiple of 4 (or your equivalent conditions factorial, e.g. all combinations length)

UnfoldSim.predef_eeg — Method

predefeeg(;kwargs...) predefeeg(rng;kwargs...) predefeeg(rng,nsubjects;kwargs...)

Gene rates a P1/N1/P3 complex. In case n_subjects is defined - MixedModelComponentsare generated, elseLinearModelComponents`

Most used kwargs is: return_epoched=true to ignore the overlap/onset bits and return already epoched data

Default params:

. n_repeats=100

tableModifyFun = x->shuffle(deepcopy(rng),x # random trial order

conditions = Dict(...),

component / signal

sfreq = 100,

p1 = (p100(;sfreq=sfreq), @formula(0~1),[5],Dict()), # P1 amp 5, no effects

n1 = (n170(;sfreq=sfreq), @formula(0~1+condition),[5,-3],Dict()), # N1 amp 5, dummycoded condition effect (levels "car", "face") of -3

p3 = (p300(;sfreq=sfreq), @formula(0~1+continuous),[5,1],Dict()), # P3 amp 5, continuous effect range [-5,5] with slope 1

noise

noiselevel = 0.2,

noise = PinkNoise(;noiselevel=noiselevel),

onset

overlap = (0.5,0.2), # offset + width/length of Uniform noise. put offset to 1 for no overlap. put width to 0 for no jitter

onset=UniformOnset(;offset=sfreq0.5overlap[1],width=sfreq0.5overlap[2]),

UnfoldSim.simulate — Method

by default call simulate with ::Abstractcomponent,::AbstractDesign`, but allow for custom types

making use of other information in simulation

UnfoldSim.simulate — Method

simulate a linearModel

julia> c = UnfoldSim.LinearModelComponent([0,1,1,0],@formula(0~1+cond),[1,2],Dict()) julia> design = MultiSubjectDesign(;nsubjects=2,nitems=50,item_between=(;:cond=>["A","B"])) julia> simulate(StableRNG(1),c,design)

UnfoldSim.simulate — Method

simulate MixedModelComponent

julia> design = MultiSubjectDesign(;nsubjects=2,nitems=50,item_between=(;:cond=>["A","B"])) julia> c = UnfoldSim.MixedModelComponent([0.,1,1,0],@formula(0~1+cond+(1|subject)),[1,2],Dict(:subject=>[2],),Dict()) julia> simulate(StableRNG(1),c,design)

UnfoldSim.simulate — Method

Simulates erp data given the specified parameters

UnfoldSim.weight_σs — Method

Weights a σs Dict for MixedModels.jl by a Float64

Finally sales it by σ_lmm, as a trick to simulate noise-free LMMs

I anticipate a function function weight_σs(σs::Dict,b_σs::Dict,σ_lmm::Float64) where each σs entry can be weighted individually